requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<3.0
drf-spectacular>=0.23.1,<0.24